fix: parameters help beacon
